add 完善标题栏分割线

+++ xsuper_titlebar/src/main/java/com/xxxxls/titlebar/ITitleBarStyle.kt
@@ -87,5 +87,5 @@ interface ITitleBarStyle {
     /**
      * 分割线高度
      */
-    fun getLineHeight(): Float?
+    fun getLineHeight(): Int?
 }
